Found EmbeddedLevelNameEnc. was returning incorrect one_at_level.
ExpandInterval attempted to avoid a function call and retrieved a mask
and an interval increment (i.e. one_at_level), but returned it's own
internal "one" and not the SciDB "one", which is shifted over one
position. Compare increment_LevelToMaskDelta and
SciDBincrement_LevelToMaskDelta for details. Note STARE uses SciDB
formatted values. This error was causing SpatialRange's intersect to
appear to drop triangles.